refactor the classes to separate file and other changes due to moving danger to common
