+++ src/components/Dashboard/Typography/Info.tsx
@@ -1,14 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-// @material-ui/core components
-import { makeStyles } from '@material-ui/core/styles';
-// core components
-import styles from '../../../assets/jss/material-dashboard-react/components/typographyStyle';
-// @ts-ignore
-const useStyles = makeStyles(styles);
-// @ts-ignore
-export default function Info(props) {
-  const classes = useStyles();
+
+import { typographyClasses as classes } from '../../../assets/jss/material-dashboard-react/components/typographyStyle';
+
+export default function Info(props: { children: any }) {
   const { children } = props;
   return <div className={classes.defaultFontStyle + ' ' + classes.infoText}>{children}</div>;
 }
