+++ src/components/Dashboard/views/Dashboard/Dashboard.tsx
@@ -22,8 +22,7 @@ import GridContainer from '../../Grid/GridContainer';
 import Table from '../../Table/Table';
 import Tasks from '../../Tasks/Tasks';
 import CustomTabs from '../../CustomTabs/CustomTabs';
-import Danger from '../../Typography/Danger';
-import { Card } from '../../../common/Card/Card';
+import { Danger, Card } from '../../../common';
 import CardHeader from '../../Card/CardHeader';
 import CardIcon from '../../Card/CardIcon';
 import CardBody from '../../Card/CardBody';
