+++ src/components/Dashboard/views/Typography/Typography.tsx
@@ -8,8 +8,8 @@ import Primary from '../../Typography/Primary';
 import Info from '../../Typography/Info';
 import Success from '../../Typography/Success';
 import Warning from '../../Typography/Warning';
-import Danger from '../../Typography/Danger';
-import { Card } from '../../../common/Card/Card';
+import { Danger } from '../../../common/Typography/Danger';
+import { Card } from '../../../common';
 import CardHeader from '../../Card/CardHeader';
 import CardBody from '../../Card/CardBody';
