+++ src/components/Dashboard/views/UpgradeToPro/UpgradeToPro.tsx
@@ -7,7 +7,7 @@ import Check from '@material-ui/icons/Check';
 // core components
 import GridItem from '../../Grid/GridItem';
 import GridContainer from '../../Grid/GridContainer';
-import Danger from '../../Typography/Danger';
+import { Danger } from '../../../common/Typography/Danger';
 import Success from '../../Typography/Success';
 import Button from '../../CustomButtons/Button';
 import { Card } from '../../../common/Card/Card';
